jsp mais ca me demande de commit za3ma

+++ src/config/ConfigurationEvenement.java
@@ -2,6 +2,6 @@ package config;
 
 public class ConfigurationEvenement 
 {
-    public static final double PLUIE_DUREE = 4;
+    public static final int PLUIE_DUREE = 4;
     
 }
